Switch Woo settings to its own tab
